pyproject.toml
[project]
name = "goit-pythonweb-hw-06"
version = "0.1.0"
description = ""
authors = [
    {name = "IYuliia",email = "catnew2012@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psycopg2 (>=2.9.10,<3.0.0)",
    "faker (>=37.1.0,<38.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
alembic = "^1.15.2"

